requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
openai==1.25.0
python-dotenv==1.0.1 
pyngrok  

#boto3==1.34.113
#botocore==1.34.113

selenium==4.20.0
webdriver-manager>=4.0.1
chromedriver-autoinstaller==0.6.3

moviepy==1.0.3
opencv-python-headless==4.10.0.82
#imageio-ffmpeg==0.5.1


#accelerate==1.6.0
# huggingface-hub==0.25.2     
# transformers>=4.38.0       
# diffusers>=0.30.3           

google-cloud-texttospeech==2.15.1

pandas==2.2.2
numpy<2
pydantic==2.7.0
pydantic-settings==2.2.1

torch==2.6.0
torchao==0.11.0
transformers==4.51.3
diffusers==0.33.1
accelerate==1.7.0

imageio-ffmpeg==0.6.0
Pillow==11.2.1

boto3==1.38.18
#python-dotenv==1.1.0

huggingface-hub==0.31.2
filelock==3.18.0
requests==2.32.3
tqdm==4.67.1
#numpy==2.0.2
packaging==24.2
jinja2==3.1.6

runpod
